+++ lark/visitors.py
@@ -15,14 +15,13 @@ _R = TypeVar('_R')
 _FUNC = Callable[..., _T]
 _DECORATED = Union[_FUNC, type]
 
-class Discard(Exception):
-    """When raising the Discard exception in a transformer callback,
+class Discard:
+    """When return the Discard singleton in a transformer callback,
     that node is discarded and won't appear in the parent.
     """
     pass
 
 # Transformers
-
 
 class _Decoratable:
     "Provides support for decorating methods with @v_args"
@@ -96,7 +95,7 @@ class Transformer(_Decoratable, ABC, Generic[_T]):
                     return f.visit_wrapper(f, tree.data, children, tree.meta)
                 else:
                     return f(children)
-            except (GrammarError, Discard):
+            except GrammarError:
                 raise
             except Exception as e:
                 raise VisitError(tree.data, tree, e)
@@ -109,22 +108,22 @@ class Transformer(_Decoratable, ABC, Generic[_T]):
         else:
             try:
                 return f(token)
-            except (GrammarError, Discard):
+            except GrammarError:
                 raise
             except Exception as e:
                 raise VisitError(token.type, token, e)
 
     def _transform_children(self, children):
         for c in children:
-            try:
-                if isinstance(c, Tree):
-                    yield self._transform_tree(c)
-                elif self.__visit_tokens__ and isinstance(c, Token):
-                    yield self._call_userfunc_token(c)
-                else:
-                    yield c
-            except Discard:
-                pass
+            if isinstance(c, Tree):
+                res = self._transform_tree(c)
+            elif self.__visit_tokens__ and isinstance(c, Token):
+                res = self._call_userfunc_token(c)
+            else:
+                res = c
+
+            if res is not Discard:
+                yield res
 
     def _transform_tree(self, tree):
         children = list(self._transform_children(tree.children))
@@ -277,15 +276,15 @@ class Transformer_NonRecursive(Transformer):
                     del stack[-size:]
                 else:
                     args = []
-                try:
-                    stack.append(self._call_userfunc(x, args))
-                except Discard:
-                    pass
+
+                res = self._call_userfunc(x, args)
+                if res is not Discard:
+                    stack.append(res)
+
             elif self.__visit_tokens__ and isinstance(x, Token):
-                try:
-                    stack.append(self._call_userfunc_token(x))
-                except Discard:
-                    pass
+                res = self._call_userfunc_token(x)
+                if res is not Discard:
+                    stack.append(res)
             else:
                 stack.append(x)